Responsive Web Design Template
Here is a web template that changes size from desktop to mobile in order to fit web content on a smaller screen.  This is good to have in case users have different preferences in how they want to view content.  Hopefully this content is off to a good use to any developers out there.